+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * bjam use clang on MACOSX
 * Simplify Travis package install specification
 * Add optional yield_to arguments
+* Make decorator copyable
 
 --------------------------------------------------------------------------------